powerBasisFE added
